chore: move fixed base scalar mul solver into a submodule
